CMS-1297: Add hasOperationDateChange util file
