+++ backend/routes/api/seasons.js
@@ -35,6 +35,7 @@ import * as USER_ROLES from "../../constants/userRoles.js";
 
 // import { createFirstComeFirstServedDateRange } from "../../utils/firstComeFirstServedHelper.js";
 import propagateWinterFeeDates from "../../utils/propagateWinterFeeDates.js";
+import hasOperationDateChanges from "../../utils/hasOperationDateChanges.js";
 import checkUserRoles, {
   getRolesFromAuth,
 } from "../../utils/checkUserRoles.js";
@@ -678,103 +679,6 @@ async function saveSeasonData({
     saveDateRangeAnnuals,
     saveGateDetail,
   ]);
-}
-
-/**
- * Detects whether the request changes any Operation date ranges.
- * Checks both updated/created dateRanges and deleted dateRange IDs.
- * @param {Object} params Parameters for operation-date change detection
- * @param {number} params.seasonId Season ID being saved
- * @param {Array} params.dateRanges DateRanges from the request payload
- * @param {Array<number>} params.deletedDateRangeIds DateRange IDs marked for deletion
- * @param {Transaction} params.transaction Database transaction
- * @returns {Promise<boolean>} True when Operation dates are changed
- */
-async function hasOperationDateChanges({
-  seasonId,
-  dateRanges,
-  deletedDateRangeIds,
-  transaction,
-}) {
-  const operationDateType = await DateType.findOne({
-    attributes: ["id"],
-    where: {
-      dateTypeNumber: DATE_TYPE.OPERATION,
-    },
-    transaction,
-  });
-
-  if (!operationDateType) {
-    return false;
-  }
-
-  const operationDateTypeId = operationDateType.id;
-
-  const existingOperationRanges = await DateRange.findAll({
-    attributes: ["id", "dateTypeId", "dateableId", "startDate", "endDate"],
-    where: {
-      seasonId,
-      dateTypeId: operationDateTypeId,
-    },
-    transaction,
-  });
-
-  const existingById = new Map(
-    existingOperationRanges.map((range) => [range.id, range]),
-  );
-
-  // New operation ranges (no ID) are always a change.
-  const hasOperationCreate = (dateRanges || []).some(
-    (dateRange) =>
-      !dateRange.id && dateRange.dateTypeId === operationDateTypeId,
-  );
-
-  if (hasOperationCreate) {
-    return true;
-  }
-
-  // Updated operation ranges: same ID but changed values.
-  const hasOperationUpdate = (dateRanges || []).some((dateRange) => {
-    if (!dateRange.id) {
-      return false;
-    }
-
-    const existing = existingById.get(dateRange.id);
-
-    if (!existing) {
-      return false;
-    }
-
-    const incomingTypeId = dateRange.dateTypeId;
-
-    // Existing operation range changed to a non-operation type.
-    if (incomingTypeId !== operationDateTypeId) {
-      return true;
-    }
-
-    const incomingStart = dateRange.startDate ?? null;
-    const incomingEnd = dateRange.endDate ?? null;
-    const existingStart = existing.startDate ?? null;
-    const existingEnd = existing.endDate ?? null;
-    const incomingDateableId = dateRange.dateableId ?? existing.dateableId;
-
-    return (
-      incomingDateableId !== existing.dateableId ||
-      incomingStart !== existingStart ||
-      incomingEnd !== existingEnd
-    );
-  });
-
-  if (hasOperationUpdate) {
-    return true;
-  }
-
-  if (!deletedDateRangeIds?.length) {
-    return false;
-  }
-
-  // Deleting an existing operation range is a change.
-  return deletedDateRangeIds.some((id) => existingById.has(id));
 }
 
 // Get all form data and DateRanges for a Feature Season
